review fix for comments skip. Moved data
