+++ src/Core/Css/src/WebScrap.Css/CssProcessor.cs
@@ -1,6 +1,7 @@
 using WebScrap.Css.Data;
 using WebScrap.Css.Contracts;
 using WebScrap.Core.Tags;
+using WebScrap.Core.Tags.Data;
 using System.Collections.Immutable;
 
 namespace WebScrap.Css;
