+++ src/Core/Css/tests/WebScrap.Css.Matching.Tests/GlobalUsings.cs
@@ -1,2 +1,3 @@
 global using NUnit.Framework;
 global using WebScrap.Core.Tags;
+global using WebScrap.Core.Tags.Data;
