+++ src/Core/Tags/WebScrap.Core.Tags/Data/TagInfo.cs
@@ -1,0 +1,6 @@
+namespace WebScrap.Core.Tags.Data;
+
+/// <summary>
+/// Represents the html info about the tag.
+/// </summary>
+public record class TagInfo(string Name, ILookup<string, string> Attributes);
