--- src/Core/Tags/WebScrap.Core.Tags/TagInfo.cs
@@ -1,3 +1,0 @@
-namespace WebScrap.Core.Tags;
-
-public record class TagInfo(string Name, ILookup<string, string> Attributes);
