--- src/Core/Tags/WebScrap.Core.Tags/UnprocessedTag.cs
@@ -1,6 +1,0 @@
-namespace WebScrap.Core.Tags;
-
-public record class UnprocessedTag(
-    TagInfo TagInfo,
-    int TagOffset, 
-    int? InnerOffset);
