+++ src/Core/Tags/WebScrap.Core.Tags/Usings.cs
@@ -3,4 +3,4 @@ global using WebScrap.Core.Tags.Tools;
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("WebScrap.Core.Tags.Tests")]
-[assembly:InternalsVisibleTo("WebScrap.Tests.IntegrationTests")]
+[assembly: InternalsVisibleTo("WebScrap.Tests.IntegrationTests")]
