+++ src/Modules/WebScrap.Modules.Extracting.Html/WebScrap.Modules.Extracting.Html/Tables/Processors/TableValuesProcessor.cs
@@ -1,4 +1,5 @@
 using WebScrap.Core.Tags;
+using WebScrap.Core.Tags.Data;
 
 namespace WebScrap.Modules.Extracting.Html.Tables;
